+++ lib/Backend/PdfTk.php
@@ -20,10 +20,11 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-namespace OCA\PdfDownloader\Util;
+namespace OCA\PdfDownloader\Backend;
 
 use mikehaertl\pdftk\Pdf as PdfTkUpstream;
-use OCA\PdfDownloader\Util\PdfTkInfoFile as InfoFile;
+
+use PdfTkInfoFile as InfoFile;
 
 /**
  * Tweak vanilla php-pdftk to suit our needs.
