+++ lib/Backend/PdfTkInfoFile.php
@@ -20,7 +20,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-namespace OCA\PdfDownloader\Util;
+namespace OCA\PdfDownloader\Backend;
 
 use \Exception;
